Add files via upload
This report presents a comprehensive guide to establishing a Linux router utilizing 
VirtualBox alongside three Ubuntu virtual machines. The primary function of the router 
is to serve as a bridge facilitating communication between the client and server virtual 
machines. While internet connectivity might not be established within this setup, the 
focus remains on fostering communication among the networked machines. Alongside 
the detailed steps for configuration, the report emphasizes transparency by providing 
information on the tools utilized and their sourcing, ensuring clarity and reproducibility 
of the setup process. Despite the absence of direct internet access, the successful 
transmission of packets between the interconnected virtual machines underscores the 
efficacy of the configured network environment.